requirements.txt
boto3==1.28.0
celery==5.3.1
dask==2023.5.0
h5py==3.9.0
pymongo==4.4.1
psycopg2-binary
redis==4.6.0
requests==2.31.0
seaborn==0.11.2
sentry-sdk==1.3.1
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.0
xgboost==2.0.3
pyarrow==12.0.1
xarray==2023.7.0
ipython == 8.14.0